Button Updates
I have changed the look and feel of the buttons to be more on brand with the rest of our user interfaces. I also changed all button icons to font awesome icons for the time being for the sake of convenience. I think I may change the icons to heroicons for things like the sun and moon and continue using font awesome for brands.